[changelog] Remove unused code
The code removed here was already moved to other
modules.
